NGFW-13316 Parse directory size as long.  Also add exception handling.

+++ uvm/api/com/untangle/uvm/SystemManager.java
@@ -65,5 +65,5 @@ public interface SystemManager
 
     void activateRadiusCertificate();
 
-    Integer getLogDirectorySize();
+    Long getLogDirectorySize();
 }
